Avoid permanently switching to foreground window keyboard layout
In e1ee48fa20, we started switching to the foreground window keyboard
layout when using some of the APIs. This was done to ensure that we'd
get the main process keyboard layout (which is what Electron windows
would use) even when the APIs are called in the render process.

This works fine if the foreground app is the same as the caller of these
APIs, but if the foreground app is something else, we are inadvertently
switching to its keyboard layout. This can happen e.g. if some Electron app
calls these APIs while it's in the background.

This goes against Windows principles — the user may choose to use
different layouts in different apps.

Ideally we should remove this code and recommend that these APIs are
only used in the main process, but for now, we can avoid breaking
changes by simply switching back to the original layout after the
relevant keyboard querying code has executed.